Code Cleanup

+++ src/main/java/slimeknights/tconstruct/smeltery/tileentity/TileTank.java
@@ -87,7 +87,7 @@ public class TileTank extends TileSmelteryComponent implements IFluidHandler {
   }
 
   public float getFluidAmountScaled() {
-    return (float) (tank.getFluid().amount - renderOffset) / (float) (tank.getCapacity() * 1.01F);
+    return (tank.getFluid().amount - renderOffset) / (tank.getCapacity() * 1.01F);
   }
 
   public boolean containsFluid() {
